# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sindra API Documentation",
        "description": "Dokumentasi API untuk sistem tiket & knowledge base",
        "contact": {
            "email": "support@sindra.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Ambil semua tiket",
                "description": "Menampilkan semua tiket beserta relasi reporter, assignees, dan attachments",
                "operationId": "ddc3a5e66a83482e874c84e71ebc4505",
                "responses": {
                    "200": {
                        "description": "Daftar tiket berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ticket"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Buat tiket baru",
                "description": "Membuat tiket pengaduan baru lengkap dengan assignee dan attachment",
                "operationId": "2a7d76295d0f9f3e0e63adb2ed4a0596",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "asset_uuid",
                                    "reporter_id",
                                    "title"
                                ],
                                "properties": {
                                    "asset_uuid": {
                                        "type": "string",
                                        "example": "b2b1a570-4e24-4ac7-a1bb-4e58a6c9471d"
                                    },
                                    "reporter_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Printer tidak menyala"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Printer Epson rusak di ruang 201"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "open",
                                            "in_progress",
                                            "resolved",
                                            "closed"
                                        ],
                                        "example": "open"
                                    },
                                    "assignees[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "attachments[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tiket berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/tickets/{id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Detail tiket",
                "description": "Menampilkan detail lengkap tiket berdasarkan ID termasuk reporter, assignees, attachments, dan discussions",
                "operationId": "da5292e25da44db69ed981a43357753c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID tiket",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail tiket berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tiket tidak ditemukan"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Perbarui tiket",
                "description": "Update judul, deskripsi, status, assignees, atau upload attachment baru",
                "operationId": "d4df6fbeb974e6e967aa95efe991e538",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID tiket yang akan diupdate",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Printer mati total"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Sudah dicek kabel dan tombol power"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "open",
                                            "in_progress",
                                            "resolved",
                                            "closed"
                                        ],
                                        "example": "in_progress"
                                    },
                                    "assignees[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 8
                                        }
                                    },
                                    "attachments[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tiket berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tiket tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Hapus tiket",
                "description": "Menghapus tiket beserta semua file attachment-nya",
                "operationId": "927b925c2e99540b159ffabb3a8c11c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID tiket yang ingin dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tiket berhasil dihapus"
                    },
                    "404": {
                        "description": "Tiket tidak ditemukan"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Ticket": {
                "title": "Ticket",
                "description": "Data tiket pengaduan",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "asset_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "example": "e7b8d2a9-1f4b-4e90-9a61-bd7e93e56a12"
                    },
                    "reporter_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "title": {
                        "type": "string",
                        "example": "Kerusakan Laptop"
                    },
                    "description": {
                        "type": "string",
                        "example": "Layar laptop pecah setelah digunakan."
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "open",
                            "in_progress",
                            "resolved",
                            "closed"
                        ],
                        "example": "open"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-15T08:30:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-15T10:00:00Z"
                    },
                    "assignees": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "user_id": {
                                    "type": "integer",
                                    "example": 3
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Budi Hartono"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "attachments": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "file_name": {
                                    "type": "string",
                                    "example": "bukti-foto.jpg"
                                },
                                "file_path": {
                                    "type": "string",
                                    "example": "storage/tickets/bukti-foto.jpg"
                                },
                                "mime_type": {
                                    "type": "string",
                                    "example": "image/jpeg"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}